requirements.txt
pandas
customtkinter
openpyxl
python-docx
pillow